helloworld/pyproject.toml
[tool.poetry]
name = "uitclass"
version = "0.1.0"
description = ""
authors = ["Muhammad Qasim <mohommadqasim1111@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.0"
uvicorn = {extras = ["standard"], version = "^0.27.1"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
